refactor: extract shared scan error helper to reduce duplication
