+++ backend/src/services/trivy_fs_scanner.rs
@@ -5,14 +5,14 @@
 
 use async_trait::async_trait;
 use bytes::Bytes;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use tracing::{info, warn};
 
 use crate::error::{AppError, Result};
 use crate::models::artifact::{Artifact, ArtifactMetadata};
-use crate::models::security::{RawFinding, Severity};
+use crate::models::security::RawFinding;
 use crate::services::image_scanner::TrivyReport;
-use crate::services::scanner_service::{sanitize_artifact_filename, Scanner};
+use crate::services::scanner_service::{convert_trivy_findings, fail_scan, ScanWorkspace, Scanner};
 
 /// Filesystem-based Trivy scanner for packages, libraries, and archives.
 pub struct TrivyFsScanner {
@@ -55,137 +55,6 @@ impl TrivyFsScanner {
         scannable_extensions
             .iter()
             .any(|ext| name_lower.ends_with(ext))
-    }
-
-    /// Build the workspace directory path for a given artifact.
-    fn workspace_dir(&self, artifact: &Artifact) -> PathBuf {
-        Path::new(&self.scan_workspace).join(artifact.id.to_string())
-    }
-
-    /// Prepare the scan workspace: write artifact content and extract archives.
-    async fn prepare_workspace(&self, artifact: &Artifact, content: &Bytes) -> Result<PathBuf> {
-        let workspace = self.workspace_dir(artifact);
-        tokio::fs::create_dir_all(&workspace)
-            .await
-            .map_err(|e| AppError::Internal(format!("Failed to create scan workspace: {}", e)))?;
-
-        // Use the original filename from the path (last segment) for correct extension detection,
-        // then sanitize to basename to prevent path traversal
-        let original_filename = artifact.path.rsplit('/').next().unwrap_or(&artifact.name);
-        let safe_filename = sanitize_artifact_filename(original_filename);
-        let artifact_path = workspace.join(&safe_filename);
-
-        tokio::fs::write(&artifact_path, content)
-            .await
-            .map_err(|e| {
-                AppError::Internal(format!("Failed to write artifact to workspace: {}", e))
-            })?;
-
-        // Extract archives into the workspace directory
-        if Self::is_archive(original_filename) {
-            if let Err(e) = Self::extract_archive(&artifact_path, &workspace).await {
-                warn!(
-                    "Failed to extract archive {}: {}. Scanning raw file instead.",
-                    artifact.name, e
-                );
-            }
-        }
-
-        Ok(workspace)
-    }
-
-    /// Check if the file is an extractable archive.
-    fn is_archive(name: &str) -> bool {
-        let lower = name.to_lowercase();
-        lower.ends_with(".tar.gz")
-            || lower.ends_with(".tgz")
-            || lower.ends_with(".whl")
-            || lower.ends_with(".jar")
-            || lower.ends_with(".war")
-            || lower.ends_with(".ear")
-            || lower.ends_with(".gem")
-            || lower.ends_with(".crate")
-            || lower.ends_with(".nupkg")
-            || lower.ends_with(".zip")
-            || lower.ends_with(".egg")
-    }
-
-    /// Extract an archive file into the given directory using system tools.
-    async fn extract_archive(archive_path: &Path, dest: &Path) -> Result<()> {
-        let name = archive_path
-            .file_name()
-            .unwrap_or_default()
-            .to_string_lossy()
-            .to_lowercase();
-
-        let output =
-            if name.ends_with(".tar.gz") || name.ends_with(".tgz") || name.ends_with(".crate") {
-                tokio::process::Command::new("tar")
-                    .args([
-                        "xzf",
-                        &archive_path.to_string_lossy(),
-                        "-C",
-                        &dest.to_string_lossy(),
-                    ])
-                    .output()
-                    .await
-            } else if name.ends_with(".zip")
-                || name.ends_with(".whl")
-                || name.ends_with(".jar")
-                || name.ends_with(".war")
-                || name.ends_with(".ear")
-                || name.ends_with(".nupkg")
-                || name.ends_with(".egg")
-            {
-                tokio::process::Command::new("unzip")
-                    .args([
-                        "-o",
-                        "-q",
-                        &archive_path.to_string_lossy(),
-                        "-d",
-                        &dest.to_string_lossy(),
-                    ])
-                    .output()
-                    .await
-            } else if name.ends_with(".gem") {
-                // Ruby gems are tar archives with a data.tar.gz inside
-                tokio::process::Command::new("tar")
-                    .args([
-                        "xf",
-                        &archive_path.to_string_lossy(),
-                        "-C",
-                        &dest.to_string_lossy(),
-                    ])
-                    .output()
-                    .await
-            } else {
-                return Ok(());
-            };
-
-        match output {
-            Ok(o) if o.status.success() => Ok(()),
-            Ok(o) => Err(AppError::Internal(format!(
-                "Archive extraction failed (exit {}): {}",
-                o.status,
-                String::from_utf8_lossy(&o.stderr)
-            ))),
-            Err(e) => Err(AppError::Internal(format!(
-                "Failed to execute extraction command: {}",
-                e
-            ))),
-        }
-    }
-
-    /// Clean up the scan workspace directory.
-    async fn cleanup_workspace(&self, artifact: &Artifact) {
-        let workspace = self.workspace_dir(artifact);
-        if let Err(e) = tokio::fs::remove_dir_all(&workspace).await {
-            warn!(
-                "Failed to clean up scan workspace {}: {}",
-                workspace.display(),
-                e
-            );
-        }
     }
 
     /// Attempt to scan using the Trivy CLI with server mode.
@@ -252,35 +121,6 @@ impl TrivyFsScanner {
         serde_json::from_slice(&output.stdout)
             .map_err(|e| AppError::Internal(format!("Failed to parse Trivy output: {}", e)))
     }
-
-    /// Convert Trivy report vulnerabilities into `RawFinding` values.
-    fn convert_findings(report: &TrivyReport) -> Vec<RawFinding> {
-        report
-            .results
-            .iter()
-            .flat_map(|result| {
-                result
-                    .vulnerabilities
-                    .as_deref()
-                    .unwrap_or(&[])
-                    .iter()
-                    .map(move |vuln| RawFinding {
-                        severity: Severity::from_str_loose(&vuln.severity)
-                            .unwrap_or(Severity::Info),
-                        title: vuln.title.clone().unwrap_or_else(|| {
-                            format!("{} in {}", vuln.vulnerability_id, vuln.pkg_name)
-                        }),
-                        description: vuln.description.clone(),
-                        cve_id: Some(vuln.vulnerability_id.clone()),
-                        affected_component: Some(format!("{} ({})", vuln.pkg_name, result.target)),
-                        affected_version: Some(vuln.installed_version.clone()),
-                        fixed_version: vuln.fixed_version.clone(),
-                        source: Some("trivy-filesystem".to_string()),
-                        source_url: vuln.primary_url.clone(),
-                    })
-            })
-            .collect()
-    }
 }
 
 #[async_trait]
@@ -308,10 +148,10 @@ impl Scanner for TrivyFsScanner {
             artifact.name, artifact.id
         );
 
-        // Prepare workspace with artifact content
-        let workspace = self.prepare_workspace(artifact, content).await?;
+        let workspace =
+            ScanWorkspace::prepare(&self.scan_workspace, None, artifact, content).await?;
 
-        // Try CLI with server mode first, then standalone, then degrade gracefully
+        // Try CLI with server mode first, then standalone
         let report = match self.scan_with_cli(&workspace).await {
             Ok(report) => report,
             Err(e) => {
@@ -322,18 +162,20 @@ impl Scanner for TrivyFsScanner {
                 match self.scan_with_standalone_cli(&workspace).await {
                     Ok(report) => report,
                     Err(e) => {
-                        warn!("Trivy filesystem scan failed for {}: {}", artifact.name, e);
-                        self.cleanup_workspace(artifact).await;
-                        return Err(AppError::Internal(format!(
-                            "Trivy filesystem scan failed for {}: {}",
-                            artifact.name, e
-                        )));
+                        return Err(fail_scan(
+                            "Trivy filesystem scan",
+                            artifact,
+                            &e,
+                            &self.scan_workspace,
+                            None,
+                        )
+                        .await);
                     }
                 }
             }
         };
 
-        let findings = Self::convert_findings(&report);
+        let findings = convert_trivy_findings(&report, "trivy-filesystem");
 
         info!(
             "Trivy filesystem scan complete for {}: {} vulnerabilities found",
@@ -341,8 +183,7 @@ impl Scanner for TrivyFsScanner {
             findings.len()
         );
 
-        // Clean up workspace
-        self.cleanup_workspace(artifact).await;
+        ScanWorkspace::cleanup(&self.scan_workspace, None, artifact).await;
 
         Ok(findings)
     }
@@ -351,6 +192,7 @@ impl Scanner for TrivyFsScanner {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::models::security::Severity;
 
     fn make_artifact(name: &str, content_type: &str, path: &str) -> Artifact {
         Artifact {
@@ -436,16 +278,16 @@ mod tests {
 
     #[test]
     fn test_is_archive() {
-        assert!(TrivyFsScanner::is_archive("foo.tar.gz"));
-        assert!(TrivyFsScanner::is_archive("foo.tgz"));
-        assert!(TrivyFsScanner::is_archive("foo.whl"));
-        assert!(TrivyFsScanner::is_archive("foo.jar"));
-        assert!(TrivyFsScanner::is_archive("foo.zip"));
-        assert!(TrivyFsScanner::is_archive("foo.gem"));
-        assert!(TrivyFsScanner::is_archive("foo.crate"));
-        assert!(TrivyFsScanner::is_archive("foo.nupkg"));
-        assert!(!TrivyFsScanner::is_archive("Cargo.lock"));
-        assert!(!TrivyFsScanner::is_archive("package.json"));
+        assert!(ScanWorkspace::is_archive("foo.tar.gz"));
+        assert!(ScanWorkspace::is_archive("foo.tgz"));
+        assert!(ScanWorkspace::is_archive("foo.whl"));
+        assert!(ScanWorkspace::is_archive("foo.jar"));
+        assert!(ScanWorkspace::is_archive("foo.zip"));
+        assert!(ScanWorkspace::is_archive("foo.gem"));
+        assert!(ScanWorkspace::is_archive("foo.crate"));
+        assert!(ScanWorkspace::is_archive("foo.nupkg"));
+        assert!(!ScanWorkspace::is_archive("Cargo.lock"));
+        assert!(!ScanWorkspace::is_archive("package.json"));
     }
 
     #[test]
@@ -468,7 +310,7 @@ mod tests {
             }],
         };
 
-        let findings = TrivyFsScanner::convert_findings(&report);
+        let findings = convert_trivy_findings(&report, "trivy-filesystem");
         assert_eq!(findings.len(), 1);
         assert_eq!(findings[0].severity, Severity::High);
         assert_eq!(findings[0].cve_id, Some("CVE-2023-12345".to_string()));
